Support for slack error reporting
