+++ WordpressSync/debug.js
@@ -7,5 +7,5 @@ process.env.debug = true;
 //run the indexpage async
 const indexCode = require('./index');
 (async () => {
-   return await indexCode();
+   return await indexCode({executionContext:{functionName:'debug'}});
 })();
